Refactor: extract TcpServerConnectionManagerBase class from TcpServerConnectionManager as to remove fan-out complexity from this class
