+++ hazelcast/src/main/java/com/hazelcast/internal/server/tcp/TcpServerConnectionErrorHandler.java
@@ -32,9 +32,9 @@ class TcpServerConnectionErrorHandler {
     private int faults;
     private long lastFaultTime;
 
-    TcpServerConnectionErrorHandler(TcpServerConnectionManager connectionManager, Address endPoint) {
+    TcpServerConnectionErrorHandler(ServerContext serverContext, Address endPoint) {
         this.endPoint = endPoint;
-        this.serverContext = connectionManager.getServer().getContext();
+        this.serverContext = serverContext;
         this.minInterval = serverContext.getConnectionMonitorInterval();
         this.maxFaults = serverContext.getConnectionMonitorMaxFaults();
         this.logger = serverContext.getLoggingService().getLogger(getClass());
